bci_flask_services/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-Cors==4.0.0

# Windows 桌面端内嵌窗口（WebView2）
pywebview

# 音乐生成相关依赖
torch
torchaudio
translate
audiocraft
